requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
loguru==0.7.2
pydantic==1.10.7
pydantic-settings==2.1.0
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.3
yfinance==0.2.18
redis==5.0.1
fastapi-cache2==0.1.9 